ragify/pyproject.toml
[tool.poetry]
name = "ragify"
version = "0.1.0"
description = ""
authors = ["Shardul Inamdar <99155739+sinamdar4@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
python-dotenv = "^1.0.1"
langchain = "^0.2.2"
pgvector = "^0.2.5"
psycopg2-binary = "^2.9.9"
langchain-community = "^0.2.3"
sentence-transformers = "^3.0.0"
langchain-huggingface = "^0.0.3"
pyarrow = "16.0.0"
langchain-postgres = "^0.0.6"
unstructured = "^0.14.5"
boto3 = "^1.34.127"
pandas = "^2.2.2"
openai = "^1.52.2"

[tool.poetry.group.dev.dependencies]

[tool.ruff.lint]
select = ["E", "F", "D", "C", "N"]
ignore = [
    "E501", # line-too-long
    "E402", # module-import-not-at-top-file
    "D203", # one-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "D100", # undocumented-public-module
    "D407",
]
per-file-ignores = { "tests/*" = ["D"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"